refactor(suite): fix TransactionReviewModalProps type-only circular dep & file cleanup
