--- packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/types/index.ts
@@ -1,3 +1,0 @@
-import { type SendState, type StakeState } from '@suite-common/wallet-core';
-
-export const isStakeState = (state: SendState | StakeState): state is StakeState => 'data' in state;
